Dataset Cleaning and Evaluation Tool (DECAT)
Fully functional DECAT program (.xlsm) with dataset examples. To use the Excel VBA program, simply place it in the same folder location as your target dataset (or datasets), then open the .xlsm (enable "macros" if asked) and click the button on the control panel. The program will present a list of all relevant files in the same folder as the .xlsm and ask which file you want to perform cleaning operations on. 

For ease of use/reference in Github, I have also extracted the main source code (.bas file) and the two userforms (.frx and .frm).